requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
apturl==0.5.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
dj-database-url==1.2.0
Django==4.1.7
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-restframework==0.0.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
duplicity==0.8.21
environs==9.5.0
fasteners==0.14.1
filelock==3.6.0
frozenlist==1.3.3
future==0.18.2
googletrans==3.0.0
gpg===1.16.0-unknown
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httplib2==0.20.2
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
magic-filter==1.0.9
Mako==1.1.3
MarkupSafe==2.0.1
marshmallow==3.19.0
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
natsort==8.0.2
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
packaging==23.0
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pipenv==11.9.0
platformdirs==2.5.1
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.21.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
rfc3986==1.5.0
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
systemd-python==234
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.13.0+ds
virtualenv-clone==0.3.0
wadllib==1.3.6
whitenoise==6.3.0
wxPython==4.0.7
xdg==5
xkit==0.0.0
yarl==1.8.2
zipp==1.0.0
